Update documentation for users of ts-jest
When our Jest suite grew large enough to warrant using Knapsack Pro, we
were somewhat surprised at the total increase in execution minutes (even
as our clock time shrunk due to parallelization). We implemented the
`--runInBand` recommendation and it showed some improvement, but there
was something else stretching our test runs out.

The cause was that `ts-jest` runs Typescript checks and diagnostics at
the start of every Jest run. This is normally a good thing, but just as
the effort to spin up a worker pool is wasted time and effort when
Knapsack is dividing up specs, re-running Typescript checks over and
over again isn't productive in a parallel execution scenario.

This change updates the README docs to point out this behavior and
suggest a setting to avoid it. This change reduced our total CI minutes
per CI run for Jest specs by over 60%.